+++ src/Tests/Mvp24Hours.Application.Pipe.Test/Integration/Caching/CacheResultsMiddlewareTest.cs
@@ -240,6 +240,11 @@ public class CacheResultsMiddlewareTest
             return [];
         }
 
+        public IList<string> GetKeys()
+        {
+            return [];
+        }
+
         public void SetLock()
         {
             IsLocked = true;